Moved link implementation to separate file.
